Implemented fragment for albums. Can show basic data and songs

+++ usecase-play-history/src/main/java/xyz/savvamirzoyan/musicplayer/usecaseplayhistory/PlayHistoryUseCase.kt
@@ -21,7 +21,7 @@ interface PlayHistoryUseCase {
             musicRepository.getAlbum("1262261"),
         ).map {
             LastPlayedPlaylistDomain(
-                id = it.id.toInt(),
+                id = it.id,
                 title = it.title,
                 lastPlayedTimeUTC = 0,
                 pictureUrl = it.coverPictureUrl
